improved gem command support

+++ org.kompiro.jamcircle.scripting.ui/src/org/kompiro/jamcircle/scripting/ui/internal/eclipse/ui/console/RubyScriptingConsole.java
@@ -479,7 +479,8 @@ public class RubyScriptingConsole extends TextConsole {
 	}
 
 	private void terminate() {
-		container.terminate();
+		if (container != null)
+			container.terminate();
 		try {
 			output.write(LINE_SEPARATOR);
 			output.write("terminated...");
